requirements.txt
requests==2.21.0
python-telegram-bot==12.2.0
gunicorn==19.9.0
Flask==1.1.1
python-dotenv==0.12.0
